LayoutTests/imported/w3c/web-platform-tests/css/css-grid/layout-algorithm/grid-minimum-contribution-baseline-shim-vertical-rl-expected.txt


FAIL .grid 1 assert_equals:
<div class="grid" style="grid-template-columns: minmax(auto, 0px);">
  <div class="item1" data-offset-x="-25" data-offset-y="0"></div>
  <div class="item2" data-offset-x="25" data-offset-y="50"></div>
  <div class="area" data-expected-width="75" data-expected-height="100"></div>
</div>
offsetLeft expected -25 but got 0
FAIL .grid 2 assert_equals:
<div class="grid" style="grid-template-columns: minmax(auto, 75px);">
  <div class="item1" data-offset-x="-25" data-offset-y="0"></div>
  <div class="item2" data-offset-x="25" data-offset-y="50"></div>
  <div class="area" data-expected-width="75" data-expected-height="100"></div>
</div>
offsetLeft expected -25 but got 0
FAIL .grid 3 assert_equals:
<div class="grid" style="grid-template-columns: minmax(auto, 88px);">
  <div class="item1" data-offset-x="-12" data-offset-y="0"></div>
  <div class="item2" data-offset-x="38" data-offset-y="50"></div>
  <div class="area" data-expected-width="88" data-expected-height="100"></div>
</div>
offsetLeft expected -12 but got 0
PASS .grid 4
PASS .grid 5
PASS .grid 6
